requirements.txt
python-dotenv==1.0.1
pydantic==2.9.2
pillow==11.0.0
PyAutoGUI==0.9.54
numpy==2.1.3